requirements.txt
streamlit
torch
torchvision
spacy
nltk
pandas
Pillow
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz